Revert "Community Page and Time Series Plot Refractor."
This reverts commit 5b005173689e2ba319b83e396e52538c7bdd2d8f.
