+++ litmus-portal/frontend/src/pages/Community/styles.ts
@@ -2,17 +2,26 @@ import { makeStyles } from '@material-ui/core/styles';
 
 // Community Component Styling
 const useStyles = makeStyles((theme) => ({
+  root: {
+    flexGrow: 1,
+    display: 'flex',
+    fontFamily: 'Ubuntu',
+    flexDirection: 'column',
+    justifyContent: 'space-between',
+  },
   mainHeader: {
     color: theme.palette.text.primary,
     fontSize: theme.spacing(5),
-    marginTop: theme.spacing(5),
     marginBottom: theme.spacing(6),
   },
-
+  formControl: {
+    margin: theme.spacing(0.5),
+    height: theme.spacing(6.5),
+    width: theme.spacing(38.75),
+  },
   LitmusAnalyticsBlock: {
     margin: '1rem 0',
   },
-
   LitmusAnalyticsDiv: {
     display: 'flex',
     flexDirection: 'row',
@@ -23,12 +32,10 @@ const useStyles = makeStyles((theme) => ({
       flexDirection: 'column',
     },
   },
-
   header2: {
     fontSize: theme.spacing(3),
     padding: '1rem 0',
   },
-
   cardDiv: {
     display: 'flex',
     flexDirection: 'row',
@@ -39,20 +46,17 @@ const useStyles = makeStyles((theme) => ({
       marginLeft: theme.spacing(8),
     },
   },
-
   paper: {
     flexGrow: 1,
     marginRight: theme.spacing(2),
     textAlign: 'center',
     color: theme.palette.text.secondary,
-    backgroundColor: '#FFF',
     [theme.breakpoints.down('sm')]: {
       display: 'flex',
       width: '100%',
       flexDirection: 'column',
     },
   },
-
   card: {
     backgroundColor: theme.palette.secondary.dark,
     height: theme.spacing(42.5),
@@ -66,17 +70,14 @@ const useStyles = makeStyles((theme) => ({
       flexDirection: 'column',
     },
   },
-
   cardContent: {
     marginTop: theme.spacing(4),
   },
-
   LitmusOnDev: {
     marginTop: theme.spacing(2),
     fontWeight: 'bold',
     fontSize: theme.spacing(3),
   },
-
   LitmusOnDevSpan: {
     position: 'relative',
     fontWeight: 'normal',
@@ -84,7 +85,6 @@ const useStyles = makeStyles((theme) => ({
     bottom: theme.spacing(2),
     left: theme.spacing(3),
   },
-
   devToLogo: {
     fill: 'white',
     filter:
@@ -94,7 +94,6 @@ const useStyles = makeStyles((theme) => ({
     width: theme.spacing(10),
     height: theme.spacing(6),
   },
-
   followBtn: {
     width: theme.spacing(20),
     height: theme.spacing(6),
@@ -103,15 +102,12 @@ const useStyles = makeStyles((theme) => ({
     color: theme.palette.text.primary,
     textTransform: 'none',
   },
-
   devToLink: {
     textDecoration: 'none',
   },
-
   LitmusUsedBlock: {
     margin: '2rem 0',
   },
-
   LitmusUsedDiv: {
     display: 'flex',
     flexDirection: 'row',
@@ -122,7 +118,6 @@ const useStyles = makeStyles((theme) => ({
       flexDirection: 'column',
     },
   },
-
   quickActionCard: {
     height: theme.spacing(42.5),
     width: theme.spacing(40),
